Support Quests (#155)
* wip quests
* Added real UI
* WIP quests
* Improved quests UI
* Refactor window management
* Added quest log
* Improved quest flow
* Support image in dialogs
* Fixed Ok button at dialog final stage

+++ src/client/IO/Components/Slider.h
@@ -54,7 +54,7 @@ namespace jrc
         Point<int16_t> start;
         Point<int16_t> end;
         int16_t buttonheight;
-        int16_t rowheight;
+        double rowheight;
         int16_t x;
         int16_t row;
         int16_t rowmax;
